remove a lot of useless observables, cleanup code and comments
