+++ AvaGui/ViewModels/DatTypes/CSSViewModel.cs
@@ -1,0 +1,7 @@
+using ReactiveUI;
+
+namespace AvaGui.ViewModels
+{
+	public class CSSViewModel : ReactiveObject, ILocoFileViewModel
+	{ }
+}
