+++ AvaGui/ViewModels/DatTypes/G1ViewModel.cs
@@ -1,0 +1,7 @@
+using ReactiveUI;
+
+namespace AvaGui.ViewModels
+{
+	public class G1ViewModel : ReactiveObject, ILocoFileViewModel
+	{ }
+}
