+++ AvaGui/ViewModels/DatTypes/ILocoFileViewModel.cs
@@ -1,0 +1,5 @@
+namespace AvaGui.ViewModels
+{
+	public interface ILocoFileViewModel
+	{ }
+}
