+++ AvaGui/ViewModels/DatTypes/TutorialViewModel.cs
@@ -1,0 +1,7 @@
+using ReactiveUI;
+
+namespace AvaGui.ViewModels
+{
+	public class TutorialViewModel : ReactiveObject, ILocoFileViewModel
+	{ }
+}
