+++ AvaGui/ViewModels/ObjectTypes/BuildingViewModel.cs
