+++ AvaGui/ViewModels/ObjectTypes/IObjectViewModel.cs
