+++ AvaGui/ViewModels/ObjectTypes/VehicleViewModel.cs
