+++ AvaGui/ViewModels/SubObjectTypes/StringTableViewModel.cs
